Replace ICU with iconv
Most C libraries include an implementation of iconv, including the
C library used by MinGW on Windows. This makes it easier to rely
on than ICU. Since this is primarily used by RTF for reading, the
slight slowdown that iconv has versus ICU does not matter. Most
Hunspell dictionaries are UTF-8, so the new TextCodec does not use
either ICU or iconv for spellchecking.